Removing unnecessary store/connect utilities from common that weren't used.
