+++ packages/office-ui-fabric-react/jest.config.js
@@ -1,3 +1,10 @@
 let { createConfig } = require('../../scripts/tasks/jest-resources');
+let path = require('path');
 
-module.exports = createConfig();
+const config = createConfig({
+  setupFiles: [
+    path.resolve(__dirname, 'lib/common/tests.js')
+  ]
+});
+
+module.exports = config;
