--- packages/office-ui-fabric-react/src/common/ISubscribable.ts
@@ -1,9 +1,0 @@
-import { IDisposable } from '../Utilities';
-
-export interface IEmptyFunction {
-  (): void;
-}
-
-export interface ISubscribable {
-  subscribe: (onChange?: IEmptyFunction) => IDisposable;
-}
