+++ packages/office-ui-fabric-react/src/common/tests.ts
@@ -1,21 +1,8 @@
-// Polyfills
-import 'es6-weak-map/implement';
-import 'es6-map/implement';
+/** Jest test setup file. */
 
-import { initializeIcons } from '@uifabric/icons/lib/index';
-console.log('initializing icons');
-initializeIcons('dist/');
+import { setIconOptions } from '@uifabric/styling';
 
-/**
- * This is a test entry point to help karma-webpack find all tests in the project.
- **/
-
-// Before loading modules, treat errors and warnings as test failures.
-console.error = console.warn = (warning: string) => {
-  throw new Error(warning);
-};
-
-const testContext = require.context('..', true, /.+\.test\.js?$/);
-
-testContext.keys().forEach(testContext);
-module.exports = testContext;
+// Suppress icon warnings.
+setIconOptions({
+  warnOnMissingIcons: false
+});
