+++ scripts/tasks/jest-disable-warnings.js
@@ -1,0 +1,4 @@
+// Fail on warnings.
+console.warn = (message) => {
+  throw new Error(message);
+};
